fix: don't show unavailable models in model selector
