+++ packages/@n8n/api-types/src/chat-hub.ts
@@ -214,6 +214,7 @@ export interface ChatModelMetadataDto {
 	capabilities: {
 		functionCalling: boolean;
 	};
+	available: boolean;
 }
 
 export interface ChatModelDto {
