+++ packages/frontend/editor-ui/src/features/ai/chatHub/chat.store.ts
@@ -766,6 +766,7 @@ export const useChatStore = defineStore(CHAT_STORE, () => {
 			metadata: baseModel?.metadata ?? {
 				capabilities: { functionCalling: false },
 				inputModalities: [],
+				available: true,
 			},
 		};
 		agents.value?.['custom-agent'].models.push(agent);
@@ -837,6 +838,7 @@ export const useChatStore = defineStore(CHAT_STORE, () => {
 				capabilities: {
 					functionCalling: true,
 				},
+				available: true,
 			},
 		};
 	}
