+++ packages/frontend/editor-ui/src/features/ai/chatHub/components/ModelSelector.vue
@@ -188,6 +188,7 @@ const menu = computed(() => {
 						capabilities: {
 							functionCalling: true,
 						},
+						available: true,
 					},
 				});
 			}
